- TripDetail페이지로 저장한 데이터 넘겨서 띄우기

+++ WithTrip/src/main/java/com/withtrip/mapper/TripPlanMapper.java
@@ -10,7 +10,7 @@ import com.withtrip.domain.TripPlanDTO;
 public interface TripPlanMapper {
 
 	// 여행계획 저장
-	int saveTripPlan(TripPlanDTO tripPlanDTO);
+	Long saveTripPlan(TripPlanDTO tripPlanDTO);
 
 	// 모든 여행계획 조회
     List<TripPlanDTO> getAllTripPlans();
